WIP: Create an XCFramework for SDL libraries
This does not quite work. At the very least building the static library with the argument "Static Library-iOS" does work, but it would take more changes to merge all of the appropriate simulator vs device libraries combined.

But I'm hoping that this is a good start if there is an interest in getting the XCFramework for SDL to build.